+++ packages/json-rpc-engine/src/v2/index.ts
@@ -23,6 +23,7 @@ export {
   isRequest,
   JsonRpcEngineError,
   selectHooks,
+  assertExpectedHooks,
 } from './utils';
 export type {
   Json,